+++ pnet-data-api-java/src/main/java/pnet/data/api/util/RestrictSapNumber.java
@@ -11,11 +11,27 @@ import java.util.Collection;
 public interface RestrictSapNumber<SELF extends Restrict<SELF>> extends Restrict<SELF>
 {
 
+    /**
+     * Restricts the SAP number.
+     *
+     * @param numbers the SAP number
+     * @return the request for chaining
+     * @deprecated use the restriction for companyNumber instead
+     */
+    @Deprecated
     default SELF sapNumber(String... numbers)
     {
         return restrict("sapNumber", (Object[]) numbers);
     }
 
+    /**
+     * Restricts the SAP number.
+     *
+     * @param numbers the SAP number
+     * @return the request for chaining
+     * @deprecated use the restriction for companyNumber instead
+     */
+    @Deprecated
     default SELF sapNumbers(Collection<String> numbers)
     {
         return sapNumber(numbers.toArray(new String[numbers.size()]));