+++ CHANGELOG.md
@@ -2,6 +2,7 @@ pnet-data-api 1.12.1
 ====================
 
 * Added data for impressum to companies.
+* SAP number is deprecated. Use company number instead.
 
 pnet-data-api 1.12.0
 ====================